refactor: rename operation to op
