+++ orzir-core/src/core.rs
@@ -4,7 +4,7 @@ pub mod dialect;
 pub mod interfaces;
 pub mod layout;
 pub mod mnemonic;
-pub mod operation;
+pub mod op;
 pub mod parse;
 pub mod print;
 pub mod region;
