+++ orzir-core/src/core/layout.rs
@@ -1,4 +1,4 @@
-use super::{block::Block, operation::OpObj};
+use super::{block::Block, op::OpObj};
 use crate::{
     impl_list_node,
     support::{list::List, storage::ArenaPtr},
